# EvmStack: Pop/Push of (u)int256 optimized