--- a/src/Nethermind/Nethermind.Evm.Benchmark/Nethermind.Evm.Benchmark.csproj
+++ b/src/Nethermind/Nethermind.Evm.Benchmark/Nethermind.Evm.Benchmark.csproj
@@ -6,13 +6,14 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
+      <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
     </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\Nethermind.Evm\Nethermind.Evm.csproj" />
       <ProjectReference Include="..\Nethermind.Logging\Nethermind.Logging.csproj" />
       <ProjectReference Include="..\Nethermind.State\Nethermind.State.csproj" />
+      <ProjectReference Include="..\..\int256\src\Nethermind.Int256\Nethermind.Int256.csproj" />
     </ItemGroup>
 
     <ItemGroup>