This PR optimizes the way (u)int256 is popped and pushed on the `EvmStack`. It uses the knowledge about the length of the obtained slice of the stack and reads/writes parts of `uint256` so that `uint256` does not have to re-evaluate this information again.

### Remarks

I found that `Benchmarks.sln` wasn't working and did some quick fixes to make the `Evm.Benchmarks` runnable. I changed how they are shaped a bit as well. This might require an additional review from a person that knows benchmarks should be shaped.

Currently, the benchmark used for this setup consists of the following steps:

1. create `EvmStack`
2. push
3. pop & return the value

In a run where the stack is created once, while values are repeatedly pushed&popped, the improvement may be even a bit bigger.

### Setup

``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1316 (1909/November2018Update/19H2)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.103
  [Host]        : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  .NET Core 5.0 : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Job=.NET Core 5.0  Runtime=.NET Core 5.0  
```

### Before

|  Method |                Value |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| **Uint256** |                    **0** | **39.79 ns** | **0.767 ns** | **1.024 ns** |     **-** |     **-** |     **-** |         **-** |
|  Int256 |                    0 | 40.61 ns | 0.786 ns | 0.735 ns |     - |     - |     - |         - |
| **Uint256** | **11579(...)39935 [78]** | **37.89 ns** | **0.552 ns** | **0.489 ns** |     **-** |     **-** |     **-** |         **-** |
|  Int256 | 11579(...)39935 [78] | 40.67 ns | 0.817 ns | 0.764 ns |     - |     - |     - |         - |
| **Uint256** | **12512(...)57123 [24]** | **37.76 ns** | **0.474 ns** | **0.443 ns** |     **-** |     **-** |     **-** |         **-** |
|  Int256 | 12512(...)57123 [24] | 40.58 ns | 0.666 ns | 0.623 ns |     - |     - |     - |         - |

### After

|  Method |                Value |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| **Uint256** |                    **0** | **31.39 ns** | **0.393 ns** | **0.328 ns** |     **-** |     **-** |     **-** |         **-** |
|  Int256 |                    0 | 35.78 ns | 0.606 ns | 0.537 ns |     - |     - |     - |         - |
| **Uint256** | **11579(...)39935 [78]** | **29.59 ns** | **0.612 ns** | **0.897 ns** |     **-** |     **-** |     **-** |         **-** |
|  Int256 | 11579(...)39935 [78] | 35.70 ns | 0.600 ns | 0.561 ns |     - |     - |     - |         - |
| **Uint256** | **12512(...)57123 [24]** | **29.23 ns** | **0.605 ns** | **0.594 ns** |     **-** |     **-** |     **-** |         **-** |
|  Int256 | 12512(...)57123 [24] | 36.58 ns | 0.681 ns | 1.060 ns |     - |     - |     - |         - |